Modifications inserted: Corrected the way to calculate the number of (#38)
slots without and with the physical layer; Inserted the accounting of
pre and post-amplifiers in the physical layer; Just check the account
lockout checks by QoTN and QoTO in the Control Plane; Inserted the input
value of the switch insertion loss into the physical layer file;
Corrected the error of the lack of influence of the amount of protection
strip slots in the physical layer. Before, the amount of protection
strip slots was fixed to the physical layer.